+++ frontend/static/__tests__/popups/ErrorPopup-test.tsx
@@ -2,6 +2,7 @@ import { mount, ReactWrapper } from 'enzyme';
 import * as React from 'react';
 import { Provider } from 'react-redux';
 
+import { Error } from '../../popups/ErrorPopup';
 import { PopupType } from '../../redux/state/AppState';
 import { RemovableError } from '../../RemovableError';
 import reduxUtils from '../redux-test-utils';
@@ -10,7 +11,6 @@ import { buildInnerHTML } from '../test-utils';
 describe('ErrorPopup', () => {
   let result: ReactWrapper;
   beforeEach(async () => {
-    const Error = require('../../popups/ErrorPopup').Error;
     const store = reduxUtils.createDtaleStore();
     store.getState().chartData = { error: 'error test', visible: true, type: PopupType.ERROR };
     buildInnerHTML({ settings: '' }, store);